+++ src/enum.ts
@@ -4,6 +4,7 @@ export enum Commands {
   SORT_JSON_DEEP = 'sort-json.sortJSONDeep',
   SORT_JSON_REVERSE = 'sort-json.sortJSONReverse',
   SORT_JSON_DEEP_REVERSE = 'sort-json.sortJSONDeepReverse',
+  SORT_JSON_BY_CUSTOM_COMPARISON = 'sort-json.sortJSONByCustomComparison',
   SET_SORT_TYPE = 'sort-json.setSortType',
   TOGGLE_CASE_SENSITIVE = 'sort-json.toggleCaseSensitive',
   SET_CASE_SENSITIVE = 'sort-json.setCaseSensitive',